Okay, properly use the new BagRegisterWorker
