+++ bag_register/src/main/scala/uk/ac/wellcome/platform/archive/bag_register/services/BagRegisterWorker.scala
@@ -1,42 +1,44 @@
 package uk.ac.wellcome.platform.archive.bag_register.services
 
-import akka.Done
+import akka.actor.ActorSystem
+import com.amazonaws.services.sqs.AmazonSQSAsync
 import grizzled.slf4j.Logging
-import io.circe.Encoder
-import uk.ac.wellcome.json.JsonUtil._
-import uk.ac.wellcome.messaging.sqs.NotificationStream
+import io.circe.{Decoder, Encoder}
+import uk.ac.wellcome.messaging.sqsworker.alpakka.{AlpakkaSQSWorker, AlpakkaSQSWorkerConfig}
+import uk.ac.wellcome.messaging.worker.models.Result
+import uk.ac.wellcome.messaging.worker.monitoring.MonitoringClient
+import uk.ac.wellcome.platform.archive.bag_register.models.RegistrationSummary
 import uk.ac.wellcome.platform.archive.common.ingests.models.BagRequest
 import uk.ac.wellcome.platform.archive.common.ingests.services.IngestUpdater
-import uk.ac.wellcome.platform.archive.common.operation.services.{
-  DiagnosticReporter,
-  OutgoingPublisher
-}
+import uk.ac.wellcome.platform.archive.common.operation.services.{IngestStepWorker, OutgoingPublisher}
 import uk.ac.wellcome.typesafe.Runnable
 
 import scala.concurrent.{ExecutionContext, Future}
 
 class BagRegisterWorker(
-  stream: NotificationStream[BagRequest],
+  alpakkaSQSWorkerConfig: AlpakkaSQSWorkerConfig,
   ingestUpdater: IngestUpdater,
-  outgoing: OutgoingPublisher,
-  reporter: DiagnosticReporter,
+  outgoingPublisher: OutgoingPublisher,
   register: Register
-)(implicit ec: ExecutionContext)
-    extends Logging
-    with Runnable {
+)(implicit
+  actorSystem: ActorSystem,
+  ec: ExecutionContext,
+  mc: MonitoringClient,
+  sc: AmazonSQSAsync,
+  decoder: Decoder[BagRequest],
+  encoder: Encoder[BagRequest]) extends Runnable with Logging with IngestStepWorker {
 
-  def run(): Future[Done] =
-    stream.run(processMessage)
+  private val worker: AlpakkaSQSWorker[BagRequest, RegistrationSummary] =
+    AlpakkaSQSWorker[BagRequest, RegistrationSummary](alpakkaSQSWorkerConfig) {
+      processMessage
+    }
 
-  def processMessage(
-    request: BagRequest
-  )(implicit
-    enc: Encoder[BagRequest]): Future[Unit] = {
+  def processMessage(bagRequest: BagRequest): Future[Result[RegistrationSummary]] =
     for {
-      result <- register.update(request.bagLocation)
-      _ <- reporter.report(request.requestId, result)
-      _ <- ingestUpdater.send(request.requestId, result, result.summary.bagId)
-      _ <- outgoing.sendIfSuccessful(result, request)
-    } yield ()
-  }
+      registrationSummary <- register.update(bagRequest.bagLocation)
+      _ <- ingestUpdater.send(bagRequest.requestId, registrationSummary, bagId = registrationSummary.summary.bagId)
+      _ <- outgoingPublisher.sendIfSuccessful(registrationSummary, bagRequest)
+    } yield toResult(registrationSummary)
+
+  override def run(): Future[Any] = worker.start
 }
